// api/generated/model/inAppNotificationResultDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */

export interface InAppNotificationResultDto {
  /** @nullable */
  _id?: string | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  redirectUrl?: string | null;
  isRead?: boolean;
  /** @nullable */
  type?: string | null;
  isSystemNotification?: boolean;
  /** @nullable */
  senderUserId?: string | null;
  createdAt?: string;
}